Format adjustment

+++ homework/20-10-22/sorting.c
@@ -1,30 +1,29 @@
-// Tan Haoqiang page 107 task 4 modified
+/* sorting.c -- Tan Haoqiang page 107 task 4 modified */
 
 #include <stdio.h>
-
-// Generally NUM <= 26, otherwise it requires rewriting
 #define NUM 3
+/* Generally NUM <= 26, otherwise it requires rewriting */
 
 int main(void)
 {
-	// Predefine
+	/* Predefine */
 	int s[NUM]; // Value
 	char t[NUM]; // Character
 
-	for (int i = 0; i < NUM; i++) t[i] = i + 'a'; // If NUM > 26, rewrite this
+	for (size_t i = 0; i < NUM; i++) t[i] = i + 'a'; // If NUM > 26, rewrite this
 
-	// Input
-	for (int i = 0; i < NUM; i++)
+	/* Input */
+	for (size_t i = 0; i < NUM; i++)
 	{
 		printf("%c = ", t[i]);
 		scanf("%d", &s[i]);
 	}
 
-	// Selection Sorting O(n^2)
-	for (int i = 0; i < NUM - 1; i++)
+	/* Selection Sorting O(n^2) */
+	for (size_t i = 0; i < NUM - 1; i++)
 	{
 		int min = i;
-		for (int j = i + 1; j < NUM; j++) if (s[min]<s[j]) min = j;
+		for (size_t j = i + 1; j < NUM; j++) if (s[min]<s[j]) min = j;
 		if (min != i)
 		{
 			s[i] ^= s[min];
@@ -37,9 +36,9 @@ int main(void)
 		}
 	}
 
-	// Output
-	for (int i = 0; i < NUM; i++) printf("%c = %d, ", t[i], s[i]);
+	/* Output */
+	for (size_t i = 0; i < NUM; i++) printf("%c = %d, ", t[i], s[i]);
 	printf("\n");
 
-	return 0; // Good habit
+	return 0;
 }
